+++ lib/grape/validations/validators/values_validator.rb
@@ -28,7 +28,7 @@ module Grape
 
           val = scrub(params[attr_name])
 
-          return if val.nil? && !required_for_root_scope?
+          return if val.nil? && !required?
           return if val != false && val.blank? && allow_blank?
           return if check_values?(val, attr_name)
 
@@ -53,15 +53,6 @@ module Grape
 
             param_array.all? { |param| values.include?(param) }
           end
-        end
-
-        def required_for_root_scope?
-          return false unless required?
-
-          current_scope = scope
-          current_scope = current_scope.parent while current_scope.lateral?
-
-          current_scope.root?
         end
       end
     end